Completed Java EE side of API (mk1)
I completed the API, version 1, for our homesec project. However, this version does not leverage a relational database, but a file system.  As of now, the MySQL connector is not working with Java EE, and I was forced to make a emulated process. That is what the TempFileDB folder in the Database Folder is for.  To utilize the files, place the HomeSec folder, from the TempFilesDB folder directly inside your C:\ drive.